generate new set id when updating sets and relocated mockdatalayer
