+++ app/src/main/java/com/app/thelocalgym/composables/TheLocalGymNavHost.kt
@@ -1,6 +1,5 @@
 package com.app.thelocalgym.composables
 
-import android.util.Range
 import android.widget.Toast
 import androidx.compose.animation.EnterTransition
 import androidx.compose.animation.ExitTransition
@@ -15,13 +14,10 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
-import com.app.thelocalgym.Exercise
-import com.app.thelocalgym.Workout
-import com.app.thelocalgym.WorkoutSet
 import com.app.thelocalgym.composables.screens.HomeScreen
 import com.app.thelocalgym.composables.screens.WorkoutDetailsScreen
 import com.app.thelocalgym.composables.screens.WorkoutsScreen
-import java.util.UUID
+import com.app.thelocalgym.repository.MockDataLayer
 
 @Composable
 fun TheLocalGymNavHost() {
@@ -99,356 +95,4 @@ private enum class Routes {
     PROGRAMS,
     WORKOUT_DETAILS,
     EXERCISES,
-}
-
-object MockDataLayer { // TODO: Create repository and mock getters and setters
-    val workouts = listOf(
-        Workout(
-            id = UUID.randomUUID().toString(),
-            name = "Push 1",
-            exercises = listOf(
-                Exercise(
-                    id = UUID.randomUUID().toString(), name = "Incline Press", sets = listOf(
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        )
-                    )
-                ),
-                Exercise(
-                    id = UUID.randomUUID().toString(), name = "Upright Row", sets = listOf(
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        )
-                    )
-                ),
-                Exercise(
-                    id = UUID.randomUUID().toString(),
-                    name = "Triceps overhead",
-                    sets = listOf(
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        )
-                    )
-                ),
-                Exercise(
-                    id = UUID.randomUUID().toString(),
-                    name = "Triceps push down",
-                    sets = listOf(
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        )
-                    )
-                ),
-                Exercise(
-                    id = UUID.randomUUID().toString(), name = "Chest press", sets = listOf(
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        )
-                    )
-                ),
-                Exercise(
-                    id = UUID.randomUUID().toString(), name = "Lat raises", sets = listOf(
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        )
-                    )
-                ),
-            )
-        ),
-        Workout(
-            id = UUID.randomUUID().toString(),
-            name = "Pull 1",
-            exercises = listOf(
-                Exercise(
-                    id = UUID.randomUUID().toString(), name = "Seated Row", sets = listOf(
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        )
-                    )
-                ),
-                Exercise(
-                    id = UUID.randomUUID().toString(), name = "Biceps curl", sets = listOf(
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        )
-                    )
-                ),
-                Exercise(
-                    id = UUID.randomUUID().toString(), name = "Lat pull down", sets = listOf(
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        )
-                    )
-                ),
-                Exercise(
-                    id = UUID.randomUUID().toString(), name = "Rear delt fly", sets = listOf(
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        )
-                    )
-                ),
-            )
-        ),
-        Workout(
-            id = UUID.randomUUID().toString(),
-            name = "Leg 1",
-            exercises = listOf(
-                Exercise(
-                    id = UUID.randomUUID().toString(), name = "Squat", sets = listOf(
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        )
-                    )
-                ),
-                Exercise(
-                    id = UUID.randomUUID().toString(), name = "Split squats", sets = listOf(
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        )
-                    )
-                ),
-                Exercise(
-                    id = UUID.randomUUID().toString(), name = "Calf raises", sets = listOf(
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        ),
-                        WorkoutSet(
-                            id = UUID.randomUUID().toString(),
-                            currentReps = 10,
-                            rpe = 8,
-                            targetReps = Range(5, 8),
-                            weight = 20
-                        )
-                    )
-                ),
-            )
-        ),
-    )
 }
